Environment organization
Removed elevator from player controller. Added environment and light
game objects to organize hierarchy and to attach a script that includes
what the script was doing elsewhere.
