+++ Assets/Echo/EchoMultiProperty.mat
@@ -52,35 +52,35 @@ Material:
       data:
         first:
           name: _Radius0
-        second: 40.0221024
+        second: 40.0597496
       data:
         first:
           name: _Radius1
-        second: 35.6254845
+        second: 40.0012436
       data:
         first:
           name: _Radius2
-        second: 40.0625687
+        second: 40.0329094
       data:
         first:
           name: _Fade0
-        second: 4.44690323
+        second: 4.45107841
       data:
         first:
           name: _Fade1
-        second: 3.95838761
+        second: 4.44458246
       data:
         first:
           name: _Fade2
-        second: 4.4513979
+        second: 4.44810486
       data:
         first:
           name: _Fade3
-        second: 4.45876312
+        second: 4.45064259
       data:
         first:
           name: _Fade4
-        second: 4.45344305
+        second: 4.44654989
       data:
         first:
           name: _Fade5
@@ -285,12 +285,12 @@ Material:
       data:
         first:
           name: _Position0
-        second: {r: -14.6780319, g: 1.3347863, b: -15.7229319, a: 0}
+        second: {r: -5.57719803, g: -3.57217253e-17, b: 8.04381657, a: 0}
       data:
         first:
           name: _Position1
-        second: {r: -18.9208565, g: 1.66454959, b: -18.1913509, a: 0}
+        second: {r: -5.32568455, g: -3.67538415e-17, b: 8.2762289, a: 0}
       data:
         first:
           name: _Position2
-        second: {r: -7.47085667, g: 1.01223254, b: -21.3643856, a: 0}
+        second: {r: -5.32568455, g: -3.67538415e-17, b: 8.2762289, a: 0}
